change file I/O API to use pre-defined origin dirs
Instead of having a hard-coded "assets" directory in the platform layer,
the filesystem API now accepts a combination of a relative file path +
an enumeration value of platform-dependent pre-defined directories.
This allows the DynApp code more flexibility for file storage depending
on storage requirements, and also restricts the DynApp to only use
directories that belong to it.