WIP C; vulkan: WIP texture; recycle linear alloc
Once all the allocations in a linear allocator are freed, the linear
allocator resets the next allocation address back to the lowest possible
address, allowing the user to re-use this memory!  Additionally, we keep
track of a list of all allocation offsets of the allocator, allowing us
to have an extra safety/metrics mechanism.

- add another checkCast
- add KORL_MEMORY_POOL remove/empty API
- complete korl_memory_allocatorLinear_reallocate such that it should
  hopefully be able to handle allocations which shrink in allocation
  page count now without crashing