remove korl_time_probe macro + more
I've abandoned the API design of having a macro which automatically
injects time probe begin/end calls around code because this is TERRIBLE
user experience for debugging.  The user cannot correctly step through
the code which the macro wraps around, so this is way too annoying for
real-world use.

also:
- add some more checkCast overloads
- fix KORL-ISSUE-000-000-051: countFormatSubstitutions: take precision
  '*' into account
- add juice to korl-commandLine --help output for fun 😎