time: report total probe count
I also did a quick profile and found that of the ~4.5 milliseconds it
takes to log a time probe report, 4 milliseconds of that time is spent
inside `korl_log_noMeta`.  With about 292 probes that comes out to each
call to log costing about 14 microseconds.  Surely we can do better than
this (in the near future, hopefully)...