WIP; fix for load state issues
The main issue being that we need a way to serialize pointers that
exist in data segments in various parts of code (like GameMemory* in
window module, or the asset database in assetCache module).

This commit begins an approach which is simply case-by-case
serialization of additional data on a per-module basis - starting with
the window module's GameMemory* value.

Consequently, we can also remove the very dumb memoryAddress<=>offset
conversion API in the memory module used to hack this functionality
originally.