WIP; fix bug & remove lulusma test font asset
The bug causing all the glyphs to render incorrectly has been fixed;
literally just a 1-line change to "invert" the glyph Y offset.

lulusma font was a pile of crap that didn't even have width for the '_'
character, so we are going to use SourceSans3 from now on by default.