WIP; add allocator_isEmpty API
And now we are using this to check to make sure that no memory leaks
have occurred in stb-ds unit tests, since that code is honestly super
suspect now that I have fixed a couple of them.  I will probably also
use this API to check for memory leaks in the future as well.

Also, in order to accomplish this, the allocator allocation enumeration
API now supports early-out escape from enumeration, and this API is
re-used for checking the allocator to see if it is empty.