stb_ds refactor; korl-file (+ simplification)
I didn't like how I was forcing myself to use endian-aware memory
packing API here, when in reality I am not ever planning to support
endian-agnostic cross-platform support for the savestate feature.
Consequently, I was able to delete a _lot_ of overly-verbose code, which
is likely a very good thing.

In addition, I made sure to look at the save_state_create time probe
with and without this commit, and these changes don't seem to introduce
any unexpected performance overhead, so I am satisfied with this for
now! 😎