WIP; add korl_gui_widgetText
Which can be used to create a text widget composed of an arbitrary
number of arrays of raw wchar_t arrays.  We can then use this to create
a widget able to display the contents of the KORL log buffer, where each
log line is completely separate from one another due to the presence of
the log line meta data tags.