fix subtle & very annoying korl-memory crash
What was happening was that the character pool used to store __FILEW__
strings in a persistent location across all code modules was changing
addresses due to being reallocated into more pages, and not having any
room in the allocator it was contained in.

The only reasonable way to solve this, as far as I can tell, is to just
create a separate allocator that _only_ contains a single allocation.
This will allow the character pool to grow if it needs to, while
ensuring that it will never move to a new address, thus invalidating
pointers to raw file name strings stored elsewhere (such as savestate
memory reports).