WIP; remove code that measures log entries
For each new log entry to korl-log, instead of measuring the log entry
first, and then sprintf into the raw log buffer after verifying size
constraints, we just print directly into the raw log at the current
offset, then assert that we didn't overflow the log or anything silly
like that.

We are also no longer calling the *_s variations of the sprintf APIs,
since those seem to be doing unnecessary things that we just don't care
about.

Also zero out the raw buffer in debug builds after testing the raw log
ring buffer, just in case it is important for any reason to have the raw
log nullified on first use (although thinking about it now, it doesn't
seem like we need to do this since we are null-terminating each new log
entry, but w/e it can't hurt).