WIP; add cull AABB to gfx_text_draw
We now only submit draw calls to the Vulkan module for the lines of text
that are included in the provided visibility AABB.

Results:

We finally did it.  We're now able to call a draw function on a cached
object with hundreds of lines of text that can potentially be
dynamically added/removed, and everything is rendering in real-time!
And it only took me like almost 2 months to figure all this shit out...