WIP; implement gfx_text_fifoRemove, fix major bug
Text can now be seamlessly cycled out of a Gfx_Text object, allowing the
user to maintain a maximum # of text lines.

The bug fixed here is due to Vulkan DeviceAssets not having their
framesSinceLastUsed property getting properly reset back to 0.  The
chosen solution involves just resetting this value every time a call is
made to _korl_vulkan_deviceAssetDatabase_get.  After briefly looking
over all the instances where this is called, it seems unlikely for this
to introduce a Vulkan device memory leak, so I'm just going to settle
with this.