WIP; add korl_gui_setLoopIndex, + more...
korl_gui_setLoopIndex allows us to create multiple widgets with the same
"identifier" in a for loop by incrementing an internal index.

This kinda resolves KORL-ISSUE-000-000-006, but probably not in the
intended way, since we aren't actually hashing the widget/window
identifier strings; we are simply combining the string pointer address
values with the korl-gui loop index to create unique hashes for the same
identifier strings when used in a for loop.

Also:
- add debug code in korl-windows-window to display the allocation counts
  for all allocators in korl-gui widgets