WIP; _HUGE_ Korl_Gfx_Text performance gains
I have refactored Korl_Gfx_Text so that each individual text line uses
only a single Korl_Resource vertex buffer to store the glyph instance
data.  This results in significantly less load on several korl modules
that have to deal with thousands of memory allocations (one for each
line) in the case that we are trying to draw huge quantities of
dynamically updated text, such as is the case with a log rendering text
gui widget.

details:
- add new API to korl-resource to give the user more control (shift
  data, update to an offset)
- fix a few critical bugs associated with korl-resource
- fix a critical bug in korl-vulkan due to shallow copy of vertex buffer
  (failing to copy vertex attribute descriptors after a resize)