WIP; mouse event restructuring
I really want to minimize the amount of code that has to go in the
OS-specific gui implementation.  I would very much prefer it if
korl-windows-gui.c was just a thin wrapper that does nothing but decode
input events into a form fit for the korl-gui module and then just pass
the event; nothing more.