WIP; strip mesh primitive vertex data from GLB
GLB files can contain many different pieces of data aside from vertex
data, as well as vertex data from other unrelated mesh primitives.
Instead of just uploading the entire binary GLB chunk into GPU memory,
we identify the specific regions of the binary chunk that make up the
vertex data of a given mesh primitive, and _only_ upload that data to a
vertex buffer device asset.

For now, it's safe to assume that we're good to just create a separate
vertex buffer for each mesh primitive, but it's easy to imagine this
being optimized by storing one giant vertex buffer that contains data
for _many_ mesh primitives.  We will consider this optimization in the
future.