enforce separation of KORL/client, add `utility`
This is part of an effort to mend a growing issue in the platform layer,
where there is a need to include a bunch of code modules in a given KORL
client module, and that code is also utilized by the platform module.
Such code, which contains _no_ platform-specific code, and is utilized
in both KORL as well as KORL clients, is hereby defined as a "utility".

In the case that a KORL module has both platform-specific code as well
as platform-agnostic code, that module will be split, and a `-utility`
version of that module will be placed in the `utility` folder of the
code directory.