WIP; add Korl_Gfx_Immediate
We want the ability to dynamically create mesh primitives on the fly
each frame, and also the ability to re-use the same dynamically
generated meshes if desired.  Korl_Gfx_Immediate allows us to achieve
this, as seen in the now stripped-down korl-gui module with the debug
drawing of scaled axis normal lines.  With the addition of an axis
normal line mesh generation helper function, this does in fact reduce
the total # of lines of code to achieve the same draw commands, with
minimum overhead (no obviously unnecessary memory copies).