WIP; add hopefully the last large refactor task
The user almost certainly _requires_ the ability to dynamically generate
a vertex buffer at run-time, and have that data persist between frames
so that it can be used in draw calls many times.  We need to make sure
the new API fully supports this, and hopefully not in a janky way!